refactor for applySimdComparison
t
