+++ velox/functions/prestosql/Comparisons.cpp
@@ -17,6 +17,7 @@
 #include "velox/functions/prestosql/Comparisons.h"
 #include <velox/common/base/Exceptions.h>
 #include "velox/functions/Udf.h"
+#include "velox/functions/lib/SIMDComparisionUtil.h"
 #include "velox/vector/BaseVector.h"
 
 namespace facebook::velox::functions {
@@ -36,71 +37,6 @@ namespace {
 /// If the vector encoding is not flat, we revert to non simd approach.
 template <typename ComparisonOp, typename Arch = xsimd::default_arch>
 struct SimdComparator {
-  template <typename T, bool isConstant>
-  inline auto loadSimdData(const T* rawData, vector_size_t offset) {
-    using d_type = xsimd::batch<T>;
-    if constexpr (isConstant) {
-      return xsimd::broadcast<T>(rawData[0]);
-    }
-    return d_type::load_unaligned(rawData + offset);
-  }
-
-  template <typename T, bool isLeftConstant, bool isRightConstant>
-  void applySimdComparison(
-      const vector_size_t begin,
-      const vector_size_t end,
-      const T* rawLhs,
-      const T* rawRhs,
-      uint8_t* rawResult) {
-    using d_type = xsimd::batch<T>;
-    constexpr auto numScalarElements = d_type::size;
-    const auto vectorEnd = (end - begin) - (end - begin) % numScalarElements;
-
-    if constexpr (numScalarElements == 2 || numScalarElements == 4) {
-      for (auto i = begin; i < vectorEnd; i += 8) {
-        rawResult[i / 8] = 0;
-        for (auto j = 0; j < 8 && (i + j) < vectorEnd; j += numScalarElements) {
-          auto left = loadSimdData<T, isLeftConstant>(rawLhs, i + j);
-          auto right = loadSimdData<T, isRightConstant>(rawRhs, i + j);
-
-          uint8_t res = simd::toBitMask(ComparisonOp()(left, right));
-          rawResult[i / 8] |= res << j;
-        }
-      }
-    } else {
-      for (auto i = begin; i < vectorEnd; i += numScalarElements) {
-        auto left = loadSimdData<T, isLeftConstant>(rawLhs, i);
-        auto right = loadSimdData<T, isRightConstant>(rawRhs, i);
-
-        auto res = simd::toBitMask(ComparisonOp()(left, right));
-        if constexpr (numScalarElements == 8) {
-          rawResult[i / 8] = res;
-        } else if constexpr (numScalarElements == 16) {
-          uint16_t* addr = reinterpret_cast<uint16_t*>(rawResult + i / 8);
-          *addr = res;
-        } else if constexpr (numScalarElements == 32) {
-          uint32_t* addr = reinterpret_cast<uint32_t*>(rawResult + i / 8);
-          *addr = res;
-        } else {
-          VELOX_FAIL("Unsupported number of scalar elements");
-        }
-      }
-    }
-
-    // Evaluate remaining values.
-    for (auto i = vectorEnd; i < end; i++) {
-      if constexpr (isRightConstant && isLeftConstant) {
-        bits::setBit(rawResult, i, ComparisonOp()(rawLhs[0], rawRhs[0]));
-      } else if constexpr (isRightConstant) {
-        bits::setBit(rawResult, i, ComparisonOp()(rawLhs[i], rawRhs[0]));
-      } else if constexpr (isLeftConstant) {
-        bits::setBit(rawResult, i, ComparisonOp()(rawLhs[0], rawRhs[i]));
-      } else {
-        bits::setBit(rawResult, i, ComparisonOp()(rawLhs[i], rawRhs[i]));
-      }
-    }
-  }
-
   template <typename T>
   inline bool compare(T& l, T& r) const {
     if constexpr (std::is_floating_point_v<T>) {
@@ -167,22 +103,22 @@ struct SimdComparator {
       if (lhs.isConstantEncoding() && rhs.isConstantEncoding()) {
         auto l = lhs.asUnchecked<ConstantVector<T>>()->valueAt(0);
         auto r = rhs.asUnchecked<ConstantVector<T>>()->valueAt(0);
-        applySimdComparison<T, true, true>(
+        applySimdComparison<T, true, true, ComparisonOp>(
             rows.begin(), rows.end(), &l, &r, rawResult);
       } else if (lhs.isConstantEncoding()) {
         auto l = lhs.asUnchecked<ConstantVector<T>>()->valueAt(0);
         auto rawRhs = rhs.asUnchecked<FlatVector<T>>()->rawValues();
-        applySimdComparison<T, true, false>(
+        applySimdComparison<T, true, false, ComparisonOp>(
             rows.begin(), rows.end(), &l, rawRhs, rawResult);
       } else if (rhs.isConstantEncoding()) {
         auto rawLhs = lhs.asUnchecked<FlatVector<T>>()->rawValues();
         auto r = rhs.asUnchecked<ConstantVector<T>>()->valueAt(0);
-        applySimdComparison<T, false, true>(
+        applySimdComparison<T, false, true, ComparisonOp>(
             rows.begin(), rows.end(), rawLhs, &r, rawResult);
       } else {
         auto rawLhs = lhs.asUnchecked<FlatVector<T>>()->rawValues();
         auto rawRhs = rhs.asUnchecked<FlatVector<T>>()->rawValues();
-        applySimdComparison<T, false, false>(
+        applySimdComparison<T, false, false, ComparisonOp>(
             rows.begin(), rows.end(), rawLhs, rawRhs, rawResult);
       }
